requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
commonmark==0.9.1
idna==3.3
importlib-metadata==4.12.0
keyring==23.7.0
Pillow==9.2.0
Pygments==2.12.0
requests==2.28.1
rich==12.5.1
typing_extensions==4.3.0
urllib3==1.26.10
zipp==3.8.1
